Move test to dedicated file
